Moved methods and mod saveData and globalData to main class.
