+++ TestMod5-2023/src/Mod.ts
@@ -83,6 +83,15 @@ export default class HelloWorld extends Mod {
 
     //@EventHandler(EventBus.Audio, Music.Crux)
 
+    ////////////////////////////////////
+    // Data Storage
+    //
+
+    @Mod.saveData<Mod>()
+    public data: ISaveData;
+    @Mod.globalData<Mod>()
+    public globalData: IGlobalData;
+
 
     @EventHandler(GameScreen, "show")
     public onGameScreenVisible(): void {
@@ -115,7 +124,7 @@ export default class HelloWorld extends Mod {
         const fromAreaID = `${fromIID[0]}${fromIID[1]}${fromXID}${fromYID}`;
         const areaID = `${IID[0]}${IID[1]}${XID}${YID}`;
 
-        var result: IAreaData = Areas.getStoredAreaData(areaID, "ID");
+        var result: IAreaData = this.getStoredAreaData(areaID, "ID");
         //this.getStoredAreaData(areaID, "AreaID");
         log.info(result);
 
@@ -126,7 +135,7 @@ export default class HelloWorld extends Mod {
             area.Claimable = false;
             area.OwnedBy = player.name;
 
-            Areas.setStoredAreaData(area)
+            this.setStoredAreaData(area)
         }
 
         return;
@@ -160,7 +169,7 @@ export default class HelloWorld extends Mod {
         }
 
 
-        var area = Areas.getAreaData(areaID);
+        var area = this.getAreaData(areaID);
 
         log.info(area);
 
@@ -207,14 +216,59 @@ export default class HelloWorld extends Mod {
          *  Are chests safe from strangers
          */
 
-
-
-
-
-
-
-
         return undefined;
+    }
+
+    /**
+     * Stores data
+     * @param area Area
+     * @returns Boolean
+     */
+    public setStoredAreaData(area: Area): boolean {
+        // initializes it if it doesn't exist
+        this.getStoredAreaData(area.AreaID, "ID");
+
+        this.data.areaData = { ID: area.AreaID, AreaData: area };
+
+        console.log("setStoredAreaData");
+        console.log(area);
+
+        return true;
+    }
+
+    /**
+  * Parses global data for area by key AreaID. Returns data if found.
+  * @param areaId Area ID to get.
+  * @param key The key returned by data.
+  * @returns Area obj if found
+  */
+    public getStoredAreaData<K extends keyof IAreaData>(areaId: string, key: K): IAreaData {
+        console.log("getStoredAreaData");
+        console.log(`areaId:${areaId}; key:${key};`);
+
+        try {
+            this.data.areaData;
+        } catch (err) {
+            var areaInfo = new Area();
+
+            areaInfo.AreaID = areaId;
+            areaInfo.Claimable = true;
+            areaInfo.OwnedBy = "";
+
+            const areaData = {
+                ID: areaId,
+                AreaData: areaInfo
+            };
+            return areaData;
+        }
+
+        const areaData = this.data.areaData;
+        const data = areaData;
+
+        console.log(`areaData:${data}`);
+
+        return data;
+
     }
 
     @Register.command("CheckArea")
@@ -232,14 +286,33 @@ export default class HelloWorld extends Mod {
         var areaId = Areas.getAreaId(player, player.x, player.y);
 
         // Get area data
-        var areaData = Areas.getAreaData(areaId);
+        var areaData = this.getAreaData(areaId);
 
         log.info(areaData);
 
 
     }
 
+    /**
+      * Get the area details from the specified position
+      * @param islandId  Island Identifier 
+      * @param x         x pos
+      * @param y         y pos
+      * @returns         object TBD
+      */
+    public getAreaData(areaId: string): Area {
+        var iArea = this.getStoredAreaData(areaId, "ID");
+        var area = new Area();
 
+        area.AreaID = iArea.AreaData.AreaID;
+        area.Claimable = iArea.AreaData.Claimable;
+        area.OwnedBy = iArea.AreaData.OwnedBy;
+
+        console.log("getAreaDetails");
+        console.log(area);
+
+        return area;
+    }
 
     private needsUpgrade(data?: { lastVersion?: string }) {
         if (!data) {
